requirements.txt
redis==4.3.4
sentence-transformers
pypdf
pyyaml
fastapi
uvicorn
pydantic
httpx
requests
fakeredis
svgwrite
discord.py
pytest
pytest-asyncio
pytest-discord
fastmcp